fix(Utils.Fonts): fix CMapFormat4.Length short overflow for large cmap subtables
CMapFormatBase.Length was declared as `abstract short`, causing
CMapFormat4.Length to cast the computed byte count to `(short)num`.
For fonts like Arial whose Format 4 subtable exceeds 32 767 bytes the
cast overflows, producing a negative value that was then written into
the cmap subtable directory offset calculation in CmapTable.WriteData.
This caused PartialStream to receive a negative length on re-parse,
triggering ArgumentOutOfRangeException.

Fix: widen Length from `short` to `int` in the abstract base and both
concrete implementations. CMapFormat0.WriteData casts explicitly to
short (safe: 262 < 32 767). CMapFormat4.WriteData writes the length
field as UInt16 to handle subtables between 32 768 and 65 535 bytes
without signed overflow, matching the TrueType spec (uint16 field).

+++ Utils.Fonts/TTF/Tables/Cmap/CMapFormat0.cs
@@ -42,7 +42,7 @@ public class CMapFormat0 : CMapFormatBase
     /// Gets the total length (in bytes) of this CMap format.
     /// This format is always 262 bytes long.
     /// </summary>
-    public override short Length => 262;
+    public override int Length => 262;
 
     /// <summary>
     /// Gets or sets the 256-byte mapping table.
@@ -97,7 +97,7 @@ public class CMapFormat0 : CMapFormatBase
     public override void WriteData(Writer data)
     {
         data.Write<Int16>(Format);
-        data.Write<Int16>(Length);
+        data.Write<Int16>((short)Length);
         data.Write<Int16>(Language);
         data.Write<byte[]>(MapBytes);
     }
